Rename router useRouter to kebab-case use-router
